+++ tesla-api/src/main/java/com/example/company/machines/domain/port/in/CreateMachineCommand.java
@@ -7,7 +7,7 @@ public record CreateMachineCommand(
         String code,
         String status,
         String processesType,
-        Integer cycleTimeSeconds,
+        Double cycleTimeSeconds,
         LocalDate lastMaintenanceDate,
         String observations
 ) {
